prevent types than 1,0

+++ api/app/vote.php
@@ -32,7 +32,10 @@ if($conn->query("SELECT COUNT(*) FROM `votes` WHERE `voterID` = '".$loggedUserID
     $result->error=true;
     die(json_encode($result));
 }
-
+if(!$type == 1 || !$type == 0){
+    $result->error=true;
+    die(json_encode($result));
+}
 $conn->query("DELETE FROM `votes` WHERE `voterID` = '".$loggedUserID."' AND `forID` = '".$forID."' LIMIT 1;");
 $conn->query("INSERT INTO `votes`(`voterID`, `type`, `forID`) VALUES ('".$loggedUserID."','".$type."','".$forID."');");
